used Long for count queries

See merge request leos/core!908

+++ tools/repository/server/src/main/java/eu/europa/ec/leos/repository/services/PackageService.java
@@ -34,7 +34,7 @@ public interface PackageService {
 
     List<LeosDocument> findDocumentsByPackageId(final String packageId, final Set<String> categories, final boolean allVersion, boolean fetchContent);
 
-    Integer getDocumentCountByPackageName(final String packageName, Set<String> categories);
+    long getDocumentCountByPackageName(final String packageName, Set<String> categories);
 
     eu.europa.ec.leos.repository.model.Package findPackageByDocumentRef(String documentRefId) throws RepositoryException;
 }
